feat: refactor SMTC module types and improve WebGL context initialization
